Polish

+++ spring-social-twitter/src/test/java/org/springframework/social/twitter/api/impl/AbstractTwitterApiTest.java
@@ -60,7 +60,7 @@ public abstract class AbstractTwitterApiTest {
 		assertEquals("http://a3.twimg.com/profile_images/1205746571/me2_300.jpg", tweet.getProfileImageUrl());
 		assertEquals("Spring Social Showcase", tweet.getSource());
 		assertEquals(1279042701000L, tweet.getCreatedAt().getTime());
-		assertEquals(Long.valueOf(123123123123L), (Long) tweet.getInReplyToStatusId());
+		assertEquals(Long.valueOf(123123123123L), tweet.getInReplyToStatusId());
 	}
 	
 	protected void assertTimelineTweets(List<Tweet> tweets) {
